Learnt about the Link module in NextJS
1. Add <Nav /> and <Links /> component

2. The <Link /> component in NextJS fetches the page attached to a specific link, it is used to link to pages within my application. It pre-renders the page so that when I click on the link, I don't have to wait for the page to load.

3. Instead of creating each link one by one, I can create a links array that contains all the navigation links, what's the benefit of doing it this way?

Creating a links array inside the navbar component is not necessary as that can be put in another file to keep the navbar component clean and reusable. The <Nav /> component can display any link since it's getting the links from a different component, the <Links /> component can also be reusable because those links can change. This just goes to show the advantage of breaking a UI into seperate components, with the flexiblity that comes with it.

Further reference - https://chatgpt.com/c/6768302a-03b4-8010-8351-fc0921cb1fb3